moved raw collection types into new raw folder
