+++ graphql/schema.graphql
@@ -390,7 +390,6 @@ enum DocumentType {
 }
 
 """
-NEW FILE ///
 Structure to represent an edited collection. Missing certain fields and chapters in it.
 Used for sending data to the front end
 """
